Support newer specs of `asm` and `naked_functions`
With the update of corresponding RFCs, naked functions can only contain
one `asm!` with `options(noreturn)` now, and we can only use `sym` and
`const` operands in the `asm!` statement.

Currently, the only solution to this seems to be doing all the pointer
arithmetics manually (pointer
arithmetics does not work in const functions).
Also, there is a quirk in the assembler where `[rip + disp]` must be
used instead of `[disp]` in IA-64 mode.