Replace `static mut` with `RacyCell`
`static mut` is going to be phased out by the Rust community.
However, our project strongly rely on not using any Rust code to do the
magic in the trampoline code.

In order to not cause trouble to us, we use `RacyCell` to provide a
global mutable storage for function pointers.
(We do exactly the same thing in C++.)